# swagger.yaml
openapi: 3.0.0

info:
  title: Dungeon Master API
  description: Internal API Reference for Dungeon Master
  version: 1.0.0

servers:
  - url: http://localhost:5000/
    description: Local development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  # CREATE TOKEN ---------------------------------------|
  /auth/create/token:
    post:
      tags: [Auth]
      requestBody:
        name: auth
        description: Create a new token to sign
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
      responses:
        201:
          description: Created a new token

  # VERIFY SIGNATURE ---------------------------------------|
  /auth/verify/signature:
    post:
      tags: [Auth]
      requestBody:
        name: auth
        description: Verify the token signed
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                signature:
                  type: string
      responses:
        201:
          description: Signature verified

  # CREATE CONSULTATION ---------------------------------------|
  /create/consultation:
    post:
      tags: [Consultations]
      requestBody:
        name: consultation
        description: New consultation
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                contact_name:
                  type: string
                contact_email:
                  type: string
                contact_bio:
                  type: string
                contact_discord:
                  type: string
                contact_telegram:
                  type: string
                preferred_contact:
                  type: string
                project_type:
                  type: string
                project_specs:
                  type: string
                specs_link:
                  type: string
                project_desc:
                  type: string
                services_req:
                  type: array
                  items:
                    type: string
                desired_delivery:
                  type: string
                budget:
                  type: string
                delivery_priorities:
                  type: string
                additional_info:
                  type: string
                submission_type:
                  type: string
                consultation_hash:
                  type: string
                feedback:
                  type: string
                rating:
                  type: number
      responses:
        201:
          description: Created a consultation record

  # CREATE APPLICATION ---------------------------------------|
  /create/application:
    post:
      tags: [Applications]
      requestBody:
        name: application
        description: New application
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email_address:
                  type: string
                discord_handle:
                  type: string
                telegram_handle:
                  type: string
                twitter_handle:
                  type: string
                github_handle:
                  type: string
                eth_address:
                  type: string
                ens_name:
                  type: string
                introduction:
                  type: string
                learning_goals:
                  type: string
                primary_skills:
                  type: array
                  items:
                    type: string
                secondary_skills:
                  type: array
                  items:
                    type: string
                skill_type:
                  type: string
                passion:
                  type: string
                favorite_media:
                  type: string
                crypto_thrills:
                  type: string
                why_raidguild:
                  type: string
                dao_familiarity:
                  type: string
                availability:
                  type: string
                crypto_exp:
                  type: string
                comments:
                  type: string
                handbook_read:
                  type: boolean
                pledge_readiness:
                  type: boolean
                referred_by:
                  type: string
      responses:
        201:
          description: Created a application record

  # CREATE MEMBER ---------------------------------------|
  /create/member:
    post:
      tags: [Members]
      requestBody:
        name: member
        description: New member
        content:
          application/json:
            schema:
              type: object
              properties:
                legacy_id:
                  type: string
                name:
                  type: string
                email_address:
                  type: string
                discord_handle:
                  type: string
                telegram_handle:
                  type: string
                twitter_handle:
                  type: string
                github_handle:
                  type: string
                eth_address:
                  type: string
                ens_name:
                  type: string
                guild_class:
                  type: string
                primary_skills:
                  type: array
                  items:
                    type: string
                secondary_skills:
                  type: array
                  items:
                    type: string
                membership_date:
                  type: string
                is_raiding:
                  type: boolean
                championed_by:
                  type: string
                application:
                  type: string
      responses:
        201:
          description: Created a member record

  # CREATE RAID ---------------------------------------|
  /create/raid:
    post:
      tags: [Raids]
      requestBody:
        name: raid
        description: New raid
        content:
          application/json:
            schema:
              type: object
              properties:
                raid_name:
                  type: string
                status:
                  type: string
                category:
                  type: string
                cleric:
                  type: string
                roles_required:
                  type: array
                  items:
                    type: string
                raid_party:
                  type: string
                invoice_address:
                  type: string
                start_date:
                  type: string
                end_date:
                  type: string
                comments:
                  type: array
                  items:
                    type: string
                consultation:
                  type: string
                related_raids:
                  type: array
                  items:
                    type: string
                portfolio:
                  type: string
                legacy:
                  type: object
                  properties:
                    airtable_id:
                      type: string
                    escrow_index:
                      type: number
                    locker_hash:
                      type: string
      responses:
        201:
          description: Created a raid record

  # CREATE PORTFOLIO ---------------------------------------|
  /create/portfolio:
    post:
      tags: [Portfolios]
      requestBody:
        name: portfolio
        description: New portfolio
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                project_desc:
                  type: string
                category:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
                case_study:
                  type: string
                repo_link:
                  type: string
                result_link:
                  type: string
      responses:
        201:
          description: Created a portfolio record

  # CREATE RAIDPARTY ---------------------------------------|
  /create/raidparty:
    post:
      tags: [RaidParties]
      requestBody:
        name: raidparty
        description: New raidparty
        content:
          application/json:
            schema:
              type: object
              properties:
                members:
                  type: array
                  items:
                    type: string
                raid:
                  type: string
      responses:
        201:
          description: Created a raidparty record

  # CREATE COMMENT ---------------------------------------|
  /create/comment:
    post:
      tags: [Comments]
      requestBody:
        name: comment
        description: New comment
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                commented_by:
                  type: string
                commented_raid:
                  type: string
      responses:
        201:
          description: Created a comment record

  # UPDATE A SINGLE MEMBER BY ID ---------------------------------------|
  /update/member/{id}:
    patch:
      tags: [Members]
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The id of the member
      requestBody:
        name: member
        description: Update member
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email_address:
                  type: string
                discord_handle:
                  type: string
                telegram_handle:
                  type: string
                github_handle:
                  type: string
                eth_address:
                  type: string
                ens_name:
                  type: string
                guild_class:
                  type: string
                primary_skills:
                  type: array
                  items:
                    type: string
                secondary_skills:
                  type: array
                  items:
                    type: string
                membership_date:
                  type: string
                is_raiding:
                  type: boolean
                championed_by:
                  type: string
                application:
                  type: string
      responses:
        201:
          description: Updated a member

  # UPDATE A SINGLE RAID BY ID ---------------------------------------|
  /update/raid/{id}:
    patch:
      tags: [Raids]
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The id of the raid
      requestBody:
        name: raid
        description: Update raid
        content:
          application/json:
            schema:
              type: object
              properties:
                raid_name:
                  type: string
                status:
                  type: string
                category:
                  type: string
                cleric_name:
                  type: string
                roles_required:
                  type: array
                  items:
                    type: string
                raid_party:
                  type: string
                invoice_address:
                  type: string
                start_date:
                  type: string
                end_date:
                  type: string
                comments:
                  type: array
                  items:
                    type: string
                related_raids:
                  type: array
                  items:
                    type: string
                portfolio:
                  type: string
      responses:
        201:
          description: Updated a raid

  # UPDATE A SINGLE PORTFOLIO BY ID ---------------------------------------|
  /update/portfolio/{id}:
    patch:
      tags: [Portfolios]
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The id of the portfolio
      requestBody:
        name: portfolio
        description: Update portfolio
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                project_desc:
                  type: string
                category:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
                case_study:
                  type: string
                repo_link:
                  type: string
                result_link:
                  type: string
      responses:
        201:
          description: Updated a portfolio

  # UPDATE A SINGLE RAIDPARTY BY ID ---------------------------------------|
  /update/raidparty/{id}:
    patch:
      tags: [RaidParties]
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The id of the raidparty
      requestBody:
        name: raidparty
        description: Update raidparty
        content:
          application/json:
            schema:
              type: object
              properties:
                members:
                  type: array
                  items:
                    type: string
                raid:
                  type: string
      responses:
        201:
          description: Updated a raidparty

  # UPDATE A SINGLE COMMENT BY ID ---------------------------------------|
  /update/comment/{id}:
    patch:
      tags: [Comments]
      parameters:
        - in: path
          name: id
          required: true
          type: string
          description: The id of the comment
      requestBody:
        name: comment
        description: Update comment
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                commented_raid:
                  type: string
      responses:
        201:
          description: Updated a comment